requirements.txt
absl-py==0.6.1
aiogram==2.0
aiohttp==3.4.4
alabaster==0.7.12
appdirs==1.4.3
asn1crypto==0.24.0
astor==0.7.1
astropy==3.0.5
async-timeout==3.0.1
attrs==18.1.0
aurman==2.18
Automat==0.7.0
Babel==2.6.0
backcall==0.1.0
beautifulsoup4==4.6.3
bleach==3.0.2
brewer2mpl==1.4.1
bs4==0.0.1
btrfsutil==1.0.0
CacheControl==0.12.5
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
colorama==0.4.1
ConfigArgParse==0.13.0
constantly==15.1.0
cryptography==2.3.1
cssselect==1.0.3
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
distlib==0.2.8
distro==1.3.0
dj-database-url==0.5.0
Django==2.1.4
django-heroku==0.3.1
docutils==0.14
entrypoints==0.2.3
feedparser==5.2.1
fitter==1.0.8
future==0.17.1
fuzzywuzzy==0.17.0
gast==0.2.0
ggplot==0.11.5
git-review==1.27.0
goodreads==0.3.2
gunicorn==19.9.0
h5py==2.8.0
heroes-build-scrapper==1.1.7
hsaudiotag3k==1.1.3
html5lib==1.0.1
hyperlink==18.0.0
idna==2.8
imagesize==1.1.0
importlib-metadata==0.7
incremental==17.5.0
ipykernel==4.9.0
ipython==7.1.1
ipython-genutils==0.1.0
ipywidgets==7.4.2
jedi==0.13.1
jeepney==0.3.1
Jinja2==2.10
jsonlines==1.2.0
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keyring==15.0.0
kiwisolver==1.0.1
libmsym==0.2.4
lockfile==0.12.2
lxml==4.2.5
MarkupSafe==1.1.0
matplotlib==2.2.3
mistune==0.8.4
msgpack==0.5.6
multidict==4.5.2
nbconvert==5.4.0
nbformat==4.4.0
nltk==3.3
nose==1.3.7
notebook==5.7.2
numpy==1.15.4
ordered-set==3.0.2
packaging==18.0
pandas==0.23.4
pandocfilters==1.4.2
parsel==1.5.1
parso==0.3.1
path.py==11.5.0
patsy==0.5.1
pep517==0.3
pexpect==4.6.0
pickleshare==0.7.5
plotly==3.3.0
pluggy==0.8.0
praw==6.0.0
prawcore==1.0.0
progress==1.4
prometheus-client==0.4.2
prompt-toolkit==2.0.7
protobuf==3.6.1
psycopg2==2.7.6.1
ptyprocess==0.6.0
pyalpm==0.8
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycosat==0.6.3
pycparser==2.18
PyDispatcher==2.0.5
pyflakes==2.0.0
Pygments==2.3.0
PyHamcrest==1.9.0
PyJWT==1.6.4
pyOpenSSL==18.0.0
pyparsing==2.3.0
PySocks==1.6.8
python-dateutil==2.7.5
python-jsonrpc-server==0.0.2
python-language-server==0.21.4
python-Levenshtein==0.12.0
python-telegram-bot==11.1.0
pytoml==0.1.20
pytz==2018.5
pyzmq==17.1.0
queuelib==1.5.0
rauth==0.7.3
regex==2018.11.22
requests==2.20.1
retrying==1.3.3
rope==0.11.0
rpy2==2.9.5
ruamel.yaml==0.15.63
scikit-gof==0.1.3
scikit-learn==0.19.2
scikit-multilearn==0.1.0
scipy==1.1.0
Scrapy==1.5.1
seaborn==0.9.0
SecretStorage==3.1.0
Send2Trash==1.5.0
service-identity==18.1.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
slizzy==1.1.10
snowballstemmer==1.2.1
Sphinx==1.8.1
sphinx-rtd-theme==0.4.2
sphinxcontrib-websupport==1.1.0
statsmodels==0.9.0
swat==1.5.0
team==1.0
telegram==0.0.1
telepot==12.7
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
twilio==6.20.0
Twisted==18.9.0
twx.botapi==3.6.2
update-checker==0.16
urllib3==1.24.1
virtualenv==16.0.0
w3lib==1.19.0
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==4.1.2
widgetsnbextension==3.4.2
xmltodict==0.11.0
yarl==1.2.6
zope.interface==4.6.0
